requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
idna==3.7
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.2.2
whitenoise==6.9.0
